+++ backend/src/main/java/com/techsage/banking/services/interfaces/BankAccountService.java
@@ -6,6 +6,7 @@ import com.techsage.banking.models.dto.BankAccountDto;
 import com.techsage.banking.models.enums.BankAccountType;
 import org.iban4j.Iban;
 
+import java.math.*;
 import java.util.List;
 
 public interface BankAccountService {
@@ -13,6 +14,6 @@ public interface BankAccountService {
     List<BankAccountDto> findByUserAndType(User user, BankAccountType type);
     List<BankAccountDto> findByType(BankAccountType type);
     BankAccount getByIban(Iban iban);
-    BankAccount create(User user, BankAccountType bankAccountType, int absoluteMinimumBalance, Double balance);
+    BankAccount create(User user, BankAccountType bankAccountType, BigDecimal absoluteMinimumBalance, BigDecimal balance);
     BankAccount update(BankAccount bankAccount);
 }
